+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 * [#7](https://github.com/dblock/log4jna/pull/8): Upgrade to log4j2 - [@cstorey](https://github.com/cstorey).
 * [#9](https://github.com/dblock/log4jna/pull/9): Mavenized build - [@claudiow](https://github.com/claudiow).
+* [#8](https://github.com/dblock/log4jna/pull/9): Added Log4j 2 configuration - [@claudiow](https://github.com/claudiow).
+* [#11](https://github.com/dblock/log4jna/pull/19): Changed project to org.dblock.log4jna - [@claudiow](https://github.com/claudiow).
 
 1.3 (3/5/2014)
 --------------
