+++ CHANGELOG.md.bak
@@ -1,0 +1,27 @@
+2.0 (pending)
+--------------
+
+* [#7](https://github.com/dblock/log4jna/pull/8): Upgrade to log4j2 - [@cstorey](https://github.com/cstorey).
+* [#9](https://github.com/dblock/log4jna/pull/9): Mavenized build - [@claudiow](https://github.com/claudiow).
+
+1.3 (3/5/2014)
+--------------
+
+* [#3](https://github.com/dblock/log4jna/pull/3): Allow setting of an application name as log4j parameters - [@tonyniemira](https://github.com/tonyniemira).
+* [#3](https://github.com/dblock/log4jna/pull/3): Allow setting of an event and category message file - [@tonyniemira](https://github.com/tonyniemira).
+* [#3](https://github.com/dblock/log4jna/pull/3): Re-throw `Win32Exception` when failing to deregister the event source - [@tonyniemira](https://github.com/tonyniemira).
+
+1.2 (5/26/2013)
+---------------
+
+* [#2](https://github.com/dblock/log4jna/pull/2): Added ability to change the log name - [@marmstrong](https://github.com/marmstrong).
+
+1.1 (6/13/2012)
+---------------
+
+* [#1](https://github.com/dblock/log4jna/pull/1): Using error handler instead of throwing exceptions - [@kjeldahl](https://github.com/kjeldahl).
+
+1.0 (9/28/2010)
+---------------
+
+* Initial public release - [@dblock](https://github.com/dblock)
