+++ log4jna-api/bin/Readme.md
@@ -1,0 +1,2 @@
+Work in progress maven migration.
+
