+++ log4jna-api/bin/src/.gitignore
@@ -1,0 +1,1 @@
+/site/
