+++ pom.xml.bak
@@ -1,0 +1,217 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+  <groupId>org.dblock.log4jna</groupId>
+  <artifactId>log4jna</artifactId>
+  <version>2.0-SNAPSHOT</version>
+  <packaging>pom</packaging>
+  <name>Log4JNA Parent</name>
+  <url>https://github.com/dblock/log4jna</url>
+  <description>Log4jna is a library of native appenders for log4j. Unlike the native implementation(s) in Log4j, this project uses JNA and therefore does not require a native DLL in a system directory or on PATH.</description>
+  <licenses>
+    <license>
+      <name>The Apache Software License, Version 2.0</name>
+      <url>http://www.apache.org/licenses/LICENSE-2.0.txt</url>
+      <distribution>repo</distribution>
+    </license>
+  </licenses>
+  <issueManagement>
+    <url>https://github.com/dblock/log4jna/issues</url>
+    <system>github</system>
+  </issueManagement>
+  <scm>
+    <connection>scm:git:https://github.com/claudiow/log4jna.git</connection>
+    <developerConnection>scm:git:https://github.com/claudiow/log4jna.git</developerConnection>
+    <url>https://github.com/claudiow/log4jna.git</url>
+  </scm>
+  <ciManagement>
+    <system>AppVeyor</system>
+    <url>https://ci.appveyor.com/project/dblock/log4jna</url>
+  </ciManagement>
+  <developers>
+    <developer>
+      <id>dblock</id>
+      <name>Daniel Doubrovkine</name>
+      <email>dblock@dblock.org</email>
+      <timezone>america/new_york</timezone>
+      <properties>
+        <picUrl>https://avatars2.githubusercontent.com/u/542335?s=50</picUrl>
+      </properties>
+    </developer>
+    <developer>
+      <id>cstaylor</id>
+      <name>Chris Taylor</name>
+      <email>cstaylor@pacbell.net</email>
+    </developer>
+    <developer>
+      <id>jim_cakalic</id>
+      <name>Jim Cakalic</name>
+      <email>jim_cakalic@na.biomerieux.com</email>
+    </developer>
+    <developer>
+      <id>tony</id>
+      <name>Tony Niemira</name>
+      <email>tony@niemira.com</email>
+    </developer>
+    <developer>
+      <id>claudiow</id>
+      <name>Claudio Trajtenberg</name>
+      <email>claudio.trajtenberg@cgtca.ca</email>
+      <timezone>america/winnipeg</timezone>
+      <properties>
+        <picUrl>https://avatars2.githubusercontent.com/u/7267903?s=50</picUrl>
+      </properties>
+    </developer>
+  </developers>
+  <build>
+    <extensions>
+      <extension>
+        <groupId>org.apache.maven.wagon</groupId>
+        <artifactId>wagon-ssh</artifactId>
+        <version>2.7</version>
+      </extension>
+    </extensions>
+    <pluginManagement>
+      <plugins>
+        <plugin>
+          <groupId>org.apache.maven.plugins</groupId>
+          <artifactId>maven-compiler-plugin</artifactId>
+          <version>${mavencompilerplugin.version}</version>
+          <configuration>
+            <source>1.8</source>
+            <target>1.8</target>
+          </configuration>
+        </plugin>
+        <plugin>
+          <groupId>org.apache.maven.plugins</groupId>
+          <artifactId>maven-javadoc-plugin</artifactId>
+          <version>${mavenjavadocplugin.version}</version>
+        </plugin>
+        <plugin>
+          <groupId>org.apache.maven.plugins</groupId>
+          <artifactId>maven-jar-plugin</artifactId>
+          <version>3.0.2</version>
+        </plugin>
+        <plugin>
+          <groupId>org.apache.maven.plugins</groupId>
+          <artifactId>maven-source-plugin</artifactId>
+          <version>3.0.1</version>
+        </plugin>
+        <plugin>
+          <groupId>org.apache.maven.plugins</groupId>
+          <artifactId>maven-gpg-plugin</artifactId>
+          <version>1.6</version>
+        </plugin>
+      </plugins>
+    </pluginManagement>
+    <plugins>
+      <plugin>
+        <groupId>org.apache.maven.plugins</groupId>
+        <artifactId>maven-javadoc-plugin</artifactId>
+        <executions>
+          <execution>
+            <goals>
+              <goal>jar</goal>
+            </goals>
+          </execution>
+        </executions>
+      </plugin>
+      <plugin>
+        <groupId>org.apache.maven.plugins</groupId>
+        <artifactId>maven-source-plugin</artifactId>
+        <executions>
+          <execution>
+            <goals>
+              <goal>jar</goal>
+            </goals>
+          </execution>
+        </executions>
+      </plugin>
+      <plugin>
+        <groupId>org.apache.maven.plugins</groupId>
+        <artifactId>maven-gpg-plugin</artifactId>
+        <executions>
+          <execution>
+            <id>sign-artifacts</id>
+            <phase>verify</phase>
+            <goals>
+              <goal>sign</goal>
+            </goals>
+          </execution>
+        </executions>
+      </plugin>
+      <plugin>
+        <groupId>org.sonatype.plugins</groupId>
+        <artifactId>nexus-staging-maven-plugin</artifactId>
+        <version>1.6.7</version>
+        <extensions>true</extensions>
+        <configuration>
+          <serverId>osshr</serverId>
+          <nexusUrl>https://oss.sonatype.org/</nexusUrl>
+          <autoReleaseAfterClose>false</autoReleaseAfterClose>
+        </configuration>
+      </plugin>
+    </plugins>
+  </build>
+  <properties>
+    <log4j.version>2.6.1</log4j.version>
+    <jna.version>4.2.2</jna.version>
+    <junit.version>4.9</junit.version>
+    <java.version>1.8</java.version>
+    <antrunplugin.version>1.8</antrunplugin.version>
+    <mavencompilerplugin.version>3.5.1</mavencompilerplugin.version>
+    <mavenjavadocplugin.version>2.10.4</mavenjavadocplugin.version>
+    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+  </properties>
+  <dependencyManagement>
+    <dependencies>
+      <dependency>
+        <groupId>org.apache.logging.log4j</groupId>
+        <artifactId>log4j-core</artifactId>
+        <version>${log4j.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>net.java.dev.jna</groupId>
+        <artifactId>jna-platform</artifactId>
+        <version>${jna.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>junit</groupId>
+        <artifactId>junit</artifactId>
+        <version>${junit.version}</version>
+        <scope>test</scope>
+      </dependency>
+    </dependencies>
+  </dependencyManagement>
+  <dependencies>
+    <dependency>
+      <groupId>junit</groupId>
+      <artifactId>junit</artifactId>
+      <scope>test</scope>
+    </dependency>
+  </dependencies>
+  <distributionManagement>
+    <!-- We have to find a location for the site -->
+    <!-- 
+    <site>
+      <id>sites</id>
+      <name>CGTCA Project Sites</name>
+      <url>scp://httpd.cgtca.ca/var/www/sites/log4jna/</url>
+    </site>
+    -->
+    <snapshotRepository>
+      <id>osshr</id>
+      <url>https://oss.sonatype.org/content/repositories/snapshots</url>
+    </snapshotRepository>
+    <repository>
+      <id>osshr</id>
+      <url>https://oss.sonatype.org/service/local/staging/deploy/maven2</url>
+    </repository>
+  </distributionManagement>
+  <modules>
+    <module>win32dll</module>
+    <module>log4jna-api</module>
+    <module>demo</module>
+    <module>assembly</module>
+  </modules>
+</project>
